feat(mistral): docs + reasoning raw
