--- examples/ai-core/src/generate-text/mistral-reasoning.ts
@@ -1,24 +1,0 @@
-import { mistral } from '@ai-sdk/mistral';
-import { generateText } from 'ai';
-import 'dotenv/config';
-
-async function main() {
-  const result = await generateText({
-    model: mistral('magistral-small-2506'),
-    prompt: 'How many "r"s are in the word "strawberry"?',
-    temperature: 0.5, // reasoning models may ignore temperature
-    maxRetries: 0,
-  });
-
-  console.log('Reasoning:');
-  console.log(result.reasoning);
-  console.log();
-
-  console.log('Text:');
-  console.log(result.text);
-  console.log();
-
-  console.log('Warnings:', result.warnings);
-}
-
-main().catch(console.error);
