+++ packages/mistral/src/mistral-chat-language-model.ts
@@ -200,20 +200,7 @@ export class MistralChatLanguageModel implements LanguageModelV2 {
     }
 
     if (text != null && text.length > 0) {
-      if (isReasoningModel(this.modelId)) {
-        const { reasoningContent, textContent } =
-          extractReasoningAndTextContent(text);
-
-        if (reasoningContent) {
-          content.push({ type: 'reasoning', text: reasoningContent });
-        }
-
-        if (textContent) {
-          content.push({ type: 'text', text: textContent });
-        }
-      } else {
-        content.push({ type: 'text', text });
-      }
+      content.push({ type: 'text', text });
     }
 
     // tool calls:
@@ -273,11 +260,6 @@ export class MistralChatLanguageModel implements LanguageModelV2 {
     };
     let chunkNumber = 0;
     let trimLeadingSpace = false;
-
-    let accumulatedText = '';
-    let inReasoningSection = false;
-    let reasoningFinished = false;
-    const isReasoning = isReasoningModel(this.modelId);
 
     return {
       stream: response.pipeThrough(
@@ -359,60 +341,7 @@ export class MistralChatLanguageModel implements LanguageModelV2 {
                 ? textContent.trimStart()
                 : textContent;
 
-              if (isReasoning) {
-                accumulatedText += processedText;
-
-                if (
-                  !inReasoningSection &&
-                  accumulatedText.includes('<think>')
-                ) {
-                  inReasoningSection = true;
-                  const beforeThink = accumulatedText.split('<think>')[0];
-                  if (beforeThink) {
-                    controller.enqueue({ type: 'text', text: beforeThink });
-                  }
-                  const afterThink = accumulatedText.split('<think>')[1] || '';
-                  accumulatedText = afterThink;
-                }
-
-                if (inReasoningSection && !reasoningFinished) {
-                  if (accumulatedText.includes('</think>')) {
-                    const parts = accumulatedText.split('</think>');
-                    const reasoningContent = parts[0];
-                    const afterReasoning = parts[1] || '';
-
-                    if (reasoningContent) {
-                      controller.enqueue({
-                        type: 'reasoning',
-                        text: reasoningContent,
-                      });
-                    }
-
-                    controller.enqueue({ type: 'reasoning-part-finish' });
-
-                    inReasoningSection = false;
-                    reasoningFinished = true;
-
-                    if (afterReasoning) {
-                      controller.enqueue({
-                        type: 'text',
-                        text: afterReasoning,
-                      });
-                    }
-
-                    accumulatedText = '';
-                  } else {
-                    controller.enqueue({
-                      type: 'reasoning',
-                      text: processedText,
-                    });
-                  }
-                } else {
-                  controller.enqueue({ type: 'text', text: processedText });
-                }
-              } else {
-                controller.enqueue({ type: 'text', text: processedText });
-              }
+              controller.enqueue({ type: 'text', text: processedText });
 
               trimLeadingSpace = false;
             }
@@ -479,31 +408,6 @@ function extractTextContent(content: z.infer<typeof mistralContentSchema>) {
   }
 
   return textContent.length ? textContent.join('') : undefined;
-}
-
-function isReasoningModel(modelId: MistralChatModelId): boolean {
-  return (
-    modelId === 'magistral-small-2506' || modelId === 'magistral-medium-2506'
-  );
-}
-
-function extractReasoningAndTextContent(text: string): {
-  reasoningContent?: string;
-  textContent?: string;
-} {
-  const reasoningMatch = text.match(/<think>(.*?)<\/think>/s);
-
-  if (!reasoningMatch) {
-    return { textContent: text };
-  }
-
-  const reasoningContent = reasoningMatch[1];
-  const textContent = text.replace(/<think>.*?<\/think>/s, '').trim();
-
-  return {
-    reasoningContent: reasoningContent || undefined,
-    textContent: textContent || undefined,
-  };
 }
 
 const mistralContentSchema = z
